moving mocks to src folder
