+++ src/mocks/event.js
@@ -1,6 +1,6 @@
 import {nanoid} from 'nanoid';
-import {pickRandomElement, generateRandomInt, pickRandomDate, pickRandomElements} from '../src/utils/random.js';
-import {EventType} from '../src/const.js';
+import {pickRandomElement, generateRandomInt, pickRandomDate, pickRandomElements} from '../utils/random.js';
+import {EventType} from '../const.js';
 
 const maxStartDaysGap = 4;
 const maxDurationEventMinutes = 135;
