+++ src/mocks/offers.js
@@ -1,5 +1,5 @@
-import {EventType} from '../src/const.js';
-import {generateRandomInt, pickRandomElements} from '../src/utils/random.js';
+import {EventType} from '../const.js';
+import {generateRandomInt, pickRandomElements} from '../utils/random.js';
 
 const offerDetails = [`Rent a car`, `Good option`, `Perfect option`, `Ideal option`, `Improve class`, `Take a boost`];
